Second pass at pep8 standards
