+++ jwst/cube_build/cube_build_io_util.py
@@ -14,9 +14,11 @@ log.setLevel(logging.DEBUG)
 # Read in dither offset file
 # For testing this is useful but possibily this might be useful during flight if the
 # images need an additional offset applied to them
-def read_offset_file(self):
+def read_offset_file(offset_list):
 
-    f = open(self.offset_list, 'r')
+    ra_offset = []
+    dec_offset = []
+    f = open(offset_list, 'r')
     i = 0
     for line in f:
         offset_str = line.split()
@@ -24,15 +26,21 @@ def read_offset_file(self):
         ra_off = offset[0]
         dec_off = offset[1]
 
-        self.ra_offset.append(ra_off)
-        self.dec_offset.append(dec_off)
+        ra_offset.append(ra_off)
+        dec_offset.append(dec_off)
         i = i + 1
     f.close()
-
+    return ra_offset, dec_offset
 
 
 #********************************************************************************
-def read_cubepars(self, instrument_info):
+def read_cubepars(par_filename,
+                  instrument,
+                  all_channel,
+                  all_subchannel,
+                  all_grating,
+                  all_filter,
+                  instrument_info):
 #********************************************************************************
     """
     Short Summary
@@ -44,22 +52,28 @@ def read_cubepars(self, instrument_info):
 
     Parameters
     ----------
-    ptab: cube parameter reference table
-    instrument_info holds the defaults scales for each channel/subchannel
+    par_filename: cube parameter reference table
+    instrument: Either MIRI or NIRSPEC
+    all_channel: all the channels contained in input data
+    all_subchannel: all subchannels contained in input data
+    all_grating: all the gratings contained in the input data
+    all_filter: all the filters contained in the input data
+    instrument_info holds the defaults spatial scales, spectral scales, roi size,
+    weighting parametes, and min and max wavelengths for each for each band
 
     Returns
     -------
     The correct elements of instrument_info are filled in
 
     """
-    if self.instrument == 'MIRI':
-        ptab = datamodels.MiriIFUCubeParsModel(self.par_filename)
-        number_bands = len(self.all_channel)
+    if instrument == 'MIRI':
+        ptab = datamodels.MiriIFUCubeParsModel(par_filename)
+        number_bands = len(all_channel)
         # pull out the channels and subcahnnels that cover the data making up the cube
         for i in range(number_bands):
-            this_channel = self.all_channel[i]
+            this_channel = all_channel[i]
             #compare_channel = 'CH'+this_channel
-            this_sub = self.all_subchannel[i]
+            this_sub = all_subchannel[i]
             # find the table entries for this combination
             for tabdata in ptab.ifucubepars_table:
                 table_channel = tabdata['channel']
@@ -99,13 +113,13 @@ def read_cubepars(self, instrument_info):
                                               table_softrad)
 
 #        print('Done reading cubepar reference file')
-    elif self.instrument == 'NIRSPEC':
-        ptab = datamodels.NirspecIFUCubeParsModel(self.par_filename)
-        number_gratings = len(self.all_grating)
+    elif instrument == 'NIRSPEC':
+        ptab = datamodels.NirspecIFUCubeParsModel(par_filename)
+        number_gratings = len(all_grating)
 
         for i in range(number_gratings):
-            this_gwa = self.all_grating[i]
-            this_filter = self.all_filter[i]
+            this_gwa = all_grating[i]
+            this_filter = all_filter[i]
             for tabdata in ptab.ifucubepars_table:
                 table_grating = tabdata['DISPERSER']
                 table_filter = tabdata['FILTER']
@@ -169,9 +183,33 @@ def read_cubepars(self, instrument_info):
 #_______________________________________________________________________
 # Read MIRI Resolution reference file
 #********************************************************************************
-def read_resolution_file(self, instrument_info):
+def read_resolution_file(resol_filename,
+                         channel,
+                         all_channel,
+                         all_subchannel,
+                         instrument_info):
+    """
+    Short Summary
+    -------------
+    If this is MIRI data and the weighting is miripsf then read in the MIRI resolition
+    file. Read weighting parameters based on which channel and subchannel we are working
+    with. Fill in the appropriated values into instrument_info
 
-    ptab = datamodels.MiriResolutionModel(self.resol_filename)
+    Parameters
+    ----------
+    resol_filename: MIRI resolution reference table
+    channel: channels working with
+    all_channel: all the channels contained in input data
+    all_subchannel: all subchannels contained in input data
+    instrument_info holds the  MIRI psf weighting parameters
+
+    Returns
+    -------
+    The correct elements of instrument_info are filled in
+
+    """
+
+    ptab = datamodels.MiriResolutionModel(resol_filename)
     table_alpha_cutoff = ptab.psf_fwhm_alpha_table['A_CUTOFF']
     table_alpha_a_short = ptab.psf_fwhm_alpha_table['A_A_SHORT']
     table_alpha_b_short = ptab.psf_fwhm_alpha_table['A_B_SHORT']
@@ -197,12 +235,12 @@ def read_resolution_file(self, instrument_info):
                                             table_beta_a_long,
                                             table_beta_b_long)
 
-    number_bands = len(self.channel)
+    number_bands = len(channel)
 
         # pull out the channels and subcahnnels that cover the data making up the cube
     for i in range(number_bands):
-        this_channel = self.all_channel[i]
-        this_sub = self.all_subchannel[i]
+        this_channel = all_channel[i]
+        this_sub = all_subchannel[i]
         compare_band = this_channel + this_sub
         for tabdata in ptab.resolving_power_table:
             table_sub_band = tabdata['SUB_BAND']
