+++ jwst/cube_build/ifu_cube.py
@@ -4,20 +4,23 @@
 import time
 import numpy as np
 import logging
+import math
 from ..model_blender import blendmeta
 from .. import datamodels
 from ..assign_wcs import pointing
+from astropy.stats import circmean
+from astropy import units as u
 from . import cube_build_wcs_util
 from . import spaxel
 from . import cube_overlap
 from . import cube_cloud
-
+from . import coord
 
 log = logging.getLogger(__name__)
 log.setLevel(logging.DEBUG)
 
 class IFUCubeData():
-# CubeData - holds all the importatn informtion for IFU Cube Building:
+# IFUCubeData - holds all the important information for IFU Cube Building:
 # wcs, data, reference data
 
     def __init__(self,
@@ -97,11 +100,23 @@ class IFUCubeData():
 
         self.spaxel = []        # list of spaxel classes
 #********************************************************************************
-# first define the number of file names that will be used to construct this cube
-# do some checks on the IFUCube to be made
-# find the ROI size
+    def check_ifucube(self):
 
-    def setup_cube(self):
+        """
+        Short Summary
+        -------------
+        Do some quick checks that the type of cube to be produced conforms to rules
+        Find the ROI size.
+
+        Parameters
+        ----------
+        None
+
+        Returns
+        -------
+        return the ROI suze to use
+
+        """
         num1 = len(self.list_par1)
         num_files = 0
         for i in range(num1):
@@ -145,7 +160,7 @@ class IFUCubeData():
         if self.interpolation == 'pointcloud':
             log.info('Region of interest spatial, wavelength  %f %f', self.rois, self.roiw)
 #________________________________________________________________________________
-
+#********************************************************************************
 # update the output name
     def define_cubename(self):
 
@@ -179,7 +194,6 @@ class IFUCubeData():
 #________________________________________________________________________________
             elif self.instrument == 'NIRSPEC':
                 fg_name = '_'
-
                 for i in range(len(self.list_par1)):
                     fg_name = fg_name + self.list_par1[i] + '-' + self.list_par2[i]
                     if(i < self.num_bands - 1):
@@ -189,21 +203,446 @@ class IFUCubeData():
                 if self.output_type == 'single':
                     newname = self.output_name_base + fg_name + 'single_s3d.fits'
 #________________________________________________________________________________
-
         if self.output_type != 'single':
             log.info('Output Name %s', newname)
-
-#        print('*** newname ****',newname)
         return newname
-
-
-#********************************************************************************
-
 
 #********************************************************************************
     def setup_ifucube_wcs(self):
 
-        cube_build_wcs_util.setup_wcs(self)
+        """
+        Short Summary
+        -------------
+        Function to determine the min and max coordinates of the spectral
+        cube, given channel & subchannel
+
+        Parameter
+        ----------
+        self.master_table:  A table that contains the channel/subchannel or
+        filter/grating for each input file
+        self.instrument_info: Default information on the MIRI and NIRSPEC instruments.
+
+        Returns
+        -------
+        Cube Dimension Information:
+        Footprint of cube: min and max of coordinates of cube.
+        If an offset list is provided then these values are applied.
+        If the coordinate system is alpha-beta (MIRI) then min and max
+        coordinates of alpha (arc sec), beta (arc sec) and lambda (microns)
+        If the coordinate system is ra-dec then the min and max of
+        ra(degress), dec (degrees) and lambda (microns) is returned.
+        """
+#________________________________________________________________________________
+# Scale is 3 dimensions and is determined from values held in  instrument_info.GetScale
+        scale = self.determine_scale()
+        self.Cdelt1 = scale[0]
+        self.Cdelt2 = scale[1]
+        self.Cdelt3 = scale[2]
+
+        parameter1 = self.list_par1
+        parameter2 = self.list_par2
+        a_min = []
+        a_max = []
+        b_min = []
+        b_max = []
+        lambda_min = []
+        lambda_max = []
+
+        self.num_bands = len(self.list_par1)
+        log.info('Number of bands in cube  %i', self.num_bands)
+
+        for i in range(self.num_bands):
+            this_a = parameter1[i]
+            this_b = parameter2[i]
+            log.debug('Working on data  from %s,%s', this_a, this_b)
+            n = len(self.master_table.FileMap[self.instrument][this_a][this_b])
+            log.debug('number of files %d ', n)
+    # each file find the min and max a and lambda (OFFSETS NEED TO BE APPLIED TO THESE VALUES)
+            for k in range(n):
+                amin = 0.0
+                amax = 0.0
+                bmin = 0.0
+                bmax = 0.0
+                lmin = 0.0
+                lmax = 0.0
+                c1_offset = 0.0
+                c2_offset = 0.0
+                ifile = self.master_table.FileMap[self.instrument][this_a][this_b][k]
+                ioffset = len(self.master_table.FileOffset[this_a][this_b]['C1'])
+                if ioffset == n:
+                    c1_offset = self.master_table.FileOffset[this_a][this_b]['C1'][k]
+                    c2_offset = self.master_table.FileOffset[this_a][this_b]['C2'][k]
+#________________________________________________________________________________
+# Open the input data model
+# Find the footprint of the image
+
+                with datamodels.IFUImageModel(ifile) as input_model:
+                    if self.instrument == 'NIRSPEC':
+                        flag_data = 0
+                        ch_footprint = cube_build_wcs_util.find_footprint_NIRSPEC(
+                            input_model,
+                            flag_data,
+                            self.coord_system)
+                        amin, amax, bmin, bmax, lmin, lmax = ch_footprint
+#________________________________________________________________________________
+                    if self.instrument == 'MIRI':
+                        ch_footprint = cube_build_wcs_util.find_footprint_MIRI(
+                            input_model,
+                            this_a,
+                            self.instrument_info,
+                            self.coord_system)
+                        amin, amax, bmin, bmax, lmin, lmax = ch_footprint
+# If a dither offset list exists then apply the dither offsets (offsets in arc seconds)
+                    amin = amin - c1_offset / 3600.0
+                    amax = amax - c1_offset / 3600.0
+                    bmin = bmin - c2_offset / 3600.0
+                    bmax = bmax - c2_offset / 3600.0
+                    a_min.append(amin)
+                    a_max.append(amax)
+                    b_min.append(bmin)
+                    b_max.append(bmax)
+                    lambda_min.append(lmin)
+                    lambda_max.append(lmax)
+#________________________________________________________________________________
+    # done looping over files determine final size of cube
+
+        final_a_min = min(a_min)
+        final_a_max = max(a_max)
+        final_b_min = min(b_min)
+        final_b_max = max(b_max)
+        final_lambda_min = min(lambda_min)
+        final_lambda_max = max(lambda_max)
+
+        if self.wavemin is not None and self.wavemin > final_lambda_min:
+            final_lambda_min = self.wavemin
+            log.info('Changed min wavelength of cube to %f ', final_lambda_min)
+
+        if self.wavemax is not None and self.wavemax < final_lambda_max:
+            final_lambda_max = self.wavemax
+            log.info('Changed max wavelength of cube to %f ', final_lambda_max)
+#________________________________________________________________________________
+        if self.instrument == 'MIRI' and self.coord_system == 'alpha-beta':
+        #  we have a 1 to 1 mapping in beta dimension.
+            nslice = self.instrument_info.GetNSlice(parameter1[0])
+            log.info('Beta Scale %f ', self.Cdelt2)
+            self.Cdelt2 = (final_b_max - final_b_min) / nslice
+            final_b_max = final_b_min + (nslice) * self.Cdelt2
+            log.info('Changed the Beta Scale dimension so we have 1 -1 mapping between beta and slice #')
+            log.info('New Beta Scale %f ', self.Cdelt2)
+#________________________________________________________________________________
+# Test that we have data (NIRSPEC NRS2 only has IFU data for 3 configurations)
+        test_a = final_a_max - final_a_min
+        test_b = final_b_max - final_b_min
+        test_w = final_lambda_max - final_lambda_min
+        tolerance1 = 0.00001
+        tolerance2 = 0.1
+        if(test_a < tolerance1 or test_b < tolerance1 or test_w < tolerance2):
+            log.info('No Valid IFU slice data found %f %f %f ', test_a, test_b, test_w)
+#________________________________________________________________________________
+        cube_footprint = (final_a_min, final_a_max, final_b_min, final_b_max,
+                      final_lambda_min, final_lambda_max)
+#________________________________________________________________________________
+    # Based on Scaling and Min and Max values determine naxis1, naxis2, naxis3
+    # set cube CRVALs, CRPIXs and xyz coords (center  x,y,z vector spaxel centers)
+
+        if self.coord_system == 'ra-dec':
+            self.set_geometry(cube_footprint)
+        else:
+            self.set_geometryAB(cube_footprint) # local coordinate system
+        self.print_cube_geometry()
+
+#********************************************************************************
+    def determine_scale(self):
+#********************************************************************************
+        """
+        Short Summary
+        -------------
+        Determine the scale (sampling) in the 3 dimensions for the cube.
+        If the IFU cube covers more than 1 band - then use the rules to
+        define the spatial and spectral sample size to use for the cube
+        Current Rule: using the minimum
+
+        Parameters
+        ----------
+        self.instrument_info holds the defaults scales for each channel/subchannel (MIRI)
+        or Grating (NIRSPEC)
+
+        Returns
+        -------
+        scale, holding the scale for the 3 dimensions of the cube/
+
+        """
+        scale = [0, 0, 0]
+        if self.instrument == 'MIRI':
+            number_bands = len(self.list_par1)
+            min_a = 1000.00
+            min_b = 1000.00
+            min_w = 1000.00
+            for i in range(number_bands):
+                this_channel = self.list_par1[i]
+                this_sub = self.list_par2[i]
+                a_scale, b_scale, w_scale = self.instrument_info.GetScale(this_channel,
+                                                                              this_sub)
+                if a_scale < min_a:
+                    min_a = a_scale
+                if b_scale < min_b:
+                    min_b = b_scale
+                if w_scale < min_w:
+                    min_w = w_scale
+            scale = [min_a, min_b, min_w]
+
+        elif self.instrument == 'NIRSPEC':
+            number_gratings = len(self.list_par1)
+            min_a = 1000.00
+            min_b = 1000.00
+            min_w = 1000.00
+
+            for i in range(number_gratings):
+                this_gwa = self.list_par1[i]
+                this_filter = self.list_par2[i]
+                a_scale, b_scale, w_scale = self.instrument_info.GetScale(this_gwa,
+                                                                          this_filter)
+                if a_scale < min_a:
+                    min_a = a_scale
+                if b_scale < min_b:
+                    min_b = b_scale
+                if w_scale < min_w:
+                    min_w = w_scale
+            scale = [min_a, min_b, min_w]
+#________________________________________________________________________________
+# check and see if the user has set the scale or set by cfg.
+
+        a_scale = scale[0]
+        if self.scale1 != 0.0:
+            a_scale = self.scale1
+
+        b_scale = scale[1]
+        if self.scale2 != 0.0:
+            b_scale = self.scale2
+        w_scale = scale[2]
+        # temp fix for large cubes - need to change to variable wavelength scale
+        if self.scalew == 0 and self.num_bands > 6:
+            w_scale = w_scale * 2
+        if self.scalew == 0 and self.num_bands > 9:
+            w_scale = w_scale * 2
+        if self.scalew != 0.0:
+            w_scale = self.scalew
+
+        scale = [a_scale, b_scale, w_scale]
+        return scale
+#_______________________________________________________________________
+#_______________________________________________________________________
+    def set_geometry(self, footprint):
+
+        """
+        Short Summary
+        -------------
+        Based on the ra,dec and wavelength footprint set up the size of Cube in
+        the tangent plane projected coordinate system.
+
+        """
+
+        ra_min, ra_max, dec_min, dec_max, lambda_min, lambda_max = footprint # in degrees
+        dec_ave = (dec_min + dec_max) / 2.0
+
+    # we can not average ra values because of the convergence of hour angles.
+        ravalues = np.zeros(2)
+        ravalues[0] = ra_min
+        ravalues[1] = ra_max
+
+        # astropy circmean assumes angles are in radians, we have angles in degrees
+        ra_ave = circmean(ravalues * u.deg).value
+        log.info('Ra average %f12.8', ra_ave)
+
+        self.Crval1 = ra_ave
+        self.Crval2 = dec_ave
+        xi_center, eta_center = coord.radec2std(self.Crval1, self.Crval2,
+                                                ra_ave, dec_ave)
+        xi_min, eta_min = coord.radec2std(self.Crval1, self.Crval2, ra_min, dec_min)
+        xi_max, eta_max = coord.radec2std(self.Crval1, self.Crval2, ra_max, dec_max)
+#________________________________________________________________________________
+# find the CRPIX1 CRPIX2 - xi and eta centered at 0,0
+# to find location of center abs of min values is how many pixels
+        n1a = int(math.ceil(math.fabs(xi_min) / self.Cdelt1))
+        n2a = int(math.ceil(math.fabs(eta_min) / self.Cdelt2))
+
+        n1b = int(math.ceil(math.fabs(xi_max) / self.Cdelt1))
+        n2b = int(math.ceil(math.fabs(eta_max) / self.Cdelt2))
+
+        xi_min = 0.0 - (n1a * self.Cdelt1) - (self.Cdelt1 / 2.0)
+        xi_max = (n1b * self.Cdelt1) + (self.Cdelt1 / 2.0)
+
+        eta_min = 0.0 - (n2a * self.Cdelt2) - (self.Cdelt2 / 2.0)
+        eta_max = (n2b * self.Cdelt2) + (self.Cdelt2 / 2.0)
+
+        self.Crpix1 = float(n1a) + 1.0
+        self.Crpix2 = float(n2a) + 1.0
+
+        self.naxis1 = n1a + n1b
+        self.naxis2 = n2a + n2b
+
+        self.a_min = xi_min
+        self.a_max = xi_max
+        self.b_min = eta_min
+        self.b_max = eta_max
+
+# center of spaxels
+        self.xcoord = np.zeros(self.naxis1)
+        xstart = xi_min + self.Cdelt1 / 2.0
+        for i in range(self.naxis1):
+            self.xcoord[i] = xstart
+            xstart = xstart + self.Cdelt1
+
+        self.ycoord = np.zeros(self.naxis2)
+        ystart = eta_min + self.Cdelt2 / 2.0
+        for i in range(self.naxis2):
+            self.ycoord[i] = ystart
+            ystart = ystart + self.Cdelt2
+
+        ygrid = np.zeros(self.naxis2 * self.naxis1)
+        xgrid = np.zeros(self.naxis2 * self.naxis1)
+
+        k = 0
+        ystart = self.ycoord[0]
+        for i in range(self.naxis2):
+            xstart = self.xcoord[0]
+            for j in range(self.naxis1):
+                xgrid[k] = xstart
+                ygrid[k] = ystart
+                xstart = xstart + self.Cdelt1
+                k = k + 1
+            ystart = ystart + self.Cdelt2
+
+
+        self.Xcenters = xgrid
+        self.Ycenters = ygrid
+#_______________________________________________________________________
+        #set up the lambda (z) coordinate of the cube
+
+        self.lambda_min = lambda_min
+        self.lambda_max = lambda_max
+        range_lambda = self.lambda_max - self.lambda_min
+        self.naxis3 = int(math.ceil(range_lambda / self.Cdelt3))
+
+         # adjust max based on integer value of naxis3
+        lambda_center = (self.lambda_max + self.lambda_min) / 2.0
+        self.lambda_min = lambda_center - (self.naxis3 / 2.0) * self.Cdelt3
+        self.lambda_max = self.lambda_min + (self.naxis3) * self.Cdelt3
+
+        self.zcoord = np.zeros(self.naxis3)
+        self.Crval3 = self.lambda_min
+        self.Crpix3 = 1.0
+        zstart = self.lambda_min + self.Cdelt3 / 2.0
+        for i in range(self.naxis3):
+            self.zcoord[i] = zstart
+            zstart = zstart + self.Cdelt3
+#_______________________________________________________________________
+
+    def set_geometryAB(self, footprint):
+        """
+        Short Summary
+        -------------
+        Based on the alpha, beta and wavelength footprint set up the size of Cube in
+        alpha-beta space. This will be a single exposure cube - small FOV
+        assume rectangular coord system
+
+        """
+        self.a_min, self.a_max, self.b_min, self.b_max, self.lambda_min, self.lambda_max = footprint
+
+        #set up the a (x) coordinates of the cube
+        range_a = self.a_max - self.a_min
+        self.naxis1 = int(math.ceil(range_a / self.Cdelt1))
+
+        # adjust min and max based on integer value of naxis1
+        a_center = (self.a_max + self.a_min) / 2.0
+        self.a_min = a_center - (self.naxis1 / 2.0) * self.Cdelt1
+        self.a_max = a_center + (self.naxis1 / 2.0) * self.Cdelt1
+
+        self.xcoord = np.zeros(self.naxis1)
+        self.Crval1 = self.a_min
+        self.Crpix1 = 0.5
+        xstart = self.a_min + self.Cdelt1 / 2.0
+        for i in range(self.naxis1):
+            self.xcoord[i] = xstart
+            xstart = xstart + self.Cdelt1
+
+#_______________________________________________________________________
+        #set up the lambda (z) coordinate of the cube
+
+        range_lambda = self.lambda_max - self.lambda_min
+        self.naxis3 = int(math.ceil(range_lambda / self.Cdelt3))
+
+         # adjust max based on integer value of naxis3
+        lambda_center = (self.lambda_max + self.lambda_min) / 2.0
+
+        self.lambda_min = lambda_center - (self.naxis3 / 2.0) * self.Cdelt3
+        self.lambda_max = lambda_center + (self.naxis3 / 2.0) * self.Cdelt3
+
+        self.lambda_max = self.lambda_min + (self.naxis3) * self.Cdelt3
+
+        self.zcoord = np.zeros(self.naxis3)
+        self.Crval3 = self.lambda_min
+        self.Crpix3 = 1.0
+        zstart = self.lambda_min + self.Cdelt3 / 2.0
+
+        for i in range(self.naxis3):
+            self.zcoord[i] = zstart
+            zstart = zstart + self.Cdelt3
+#_______________________________________________________________________
+        # set up the naxis2 parameters
+        range_b = self.b_max - self.b_min
+
+        self.naxis2 = int(math.ceil(range_b / self.Cdelt2))
+        b_center = (self.b_max + self.b_min) / 2.0
+    # adjust min and max based on integer value of naxis2
+        self.b_max = b_center + (self.naxis2 / 2.0) * self.Cdelt2
+        self.b_min = b_center - (self.naxis2 / 2.0) * self.Cdelt2
+
+        self.ycoord = np.zeros(self.naxis2)
+        self.Crval2 = self.b_min
+        self.Crpix2 = 0.5
+        ystart = self.b_min + self.Cdelt2 / 2.0
+        for i in range(self.naxis2):
+            self.ycoord[i] = ystart
+            ystart = ystart + self.Cdelt2
+
+#_______________________________________________________________________
+    def print_cube_geometry(self):
+
+        """
+        Print out the general properties of the size of the IFU Cube
+        """
+
+        log.info('Cube Geometry:')
+        if self.coord_system == 'alpha-beta':
+            log.info('axis# Naxis  CRPIX    CRVAL      CDELT(arc sec)  MIN & Max (alpha,beta arc sec)')
+        else:
+            log.info('axis# Naxis  CRPIX    CRVAL      CDELT(arc sec)  MIN & Max (xi,eta arc sec)')
+            log.info('Axis 1 %5d  %5.2f %12.8f %12.8f %12.8f %12.8f',
+                     self.naxis1, self.Crpix1, self.Crval1, self.Cdelt1,
+                     self.a_min, self.a_max)
+            log.info('Axis 2 %5d  %5.2f %12.8f %12.8f %12.8f %12.8f',
+                     self.naxis2, self.Crpix2, self.Crval2, self.Cdelt2,
+                     self.b_min, self.b_max)
+            log.info('Axis 3 %5d  %5.2f %12.8f %12.8f %12.8f %12.8f',
+                     self.naxis3, self.Crpix3, self.Crval3, self.Cdelt3,
+                     self.lambda_min, self.lambda_max)
+
+        if self.instrument == 'MIRI':
+        # length of channel and subchannel are the same
+            number_bands = len(self.list_par1)
+            for i in range(number_bands):
+                this_channel = self.list_par1[i]
+                this_subchannel = self.list_par2[i]
+                log.info('Cube covers channel, subchannel: %s %s ', this_channel, this_subchannel)
+        elif self.instrument == 'NIRSPEC':
+            # number of filters and gratings are the same
+            number_bands = len(self.list_par1)
+            for i in range(number_bands):
+                this_fwa = self.list_par2[i]
+                this_gwa = self.list_par1[i]
+                log.info('Cube covers grating, filter: %s %s ', this_gwa, this_fwa)
+#________________________________________________________________________________
 
 #********************************************************************************
     def build_ifucube(self):
@@ -225,7 +664,6 @@ class IFUCubeData():
         """
 
         self.output_name = self.define_cubename()
-#        self.find_output_type()
         self.spaxel = self.create_spaxel()
 
         # now need to loop over every file that covers this channel/subchannel (MIRI)
@@ -253,11 +691,11 @@ class IFUCubeData():
         t1 = time.time()
         log.info("Time to find Cube Flux= %.1f.s" % (t1 - t0,))
 
-        IFUCube = self.setup_IFUCube(0)
+        ifucube_model = self.setup_ifucube(0)
 #_______________________________________________________________________
 # shove Flux and iflux in the  final IFU cube
-        self.update_IFUCube(IFUCube, self.spaxel)
-        return IFUCube
+        self.update_ifucube(ifucube_model, self.spaxel)
+        return ifucube_model
 
 #********************************************************************************
 
@@ -281,7 +719,7 @@ class IFUCubeData():
 
         # loop over input models
 
-        single_IFUCube = datamodels.ModelContainer()
+        single_ifucube_container = datamodels.ModelContainer()
         n = len(self.input_models)
         log.info("Number of Single IFU cubes creating  = %i" % n)
         this_par1 = self.list_par1[0] # only one channel is used in this approach
@@ -296,9 +734,7 @@ class IFUCubeData():
 # for each new data model create a new spaxel
             spaxel = []
             spaxel = self.create_spaxel()
-
             with datamodels.IFUImageModel(self.input_models[j]) as input_model:
-
 #********************************************************************************
 # pulled necessary routines from   CubeData.map_detector_to_spaxel
                 if self.instrument == 'MIRI':
@@ -331,21 +767,21 @@ class IFUCubeData():
                                                   c1_offset, c2_offset)
 
                         t1a = time.time()
-                        log.debug("Time Match one NIRSPEC slice  to IFUCube = %.1f.s" % (t1a - t0a,))
+                        log.debug("Time Match one NIRSPEC slice to ifucube = %.1f.s" % (t1a - t0a,))
 #_______________________________________________________________________
-# shove Flux and iflux in the  final IFU cube
+# shove Flux and iflux in the  final ifucube
 
             self.find_spaxel_flux(spaxel)
 # now determine Cube Spaxel flux
 
-            IFUCube = self.setup_IFUCube(j)
-            self.update_IFUCube(IFUCube, spaxel)
+            ifucube_model = self.setup_ifucube(j)
+            self.update_ifucube(ifucube_model, spaxel)
             t1 = time.time()
-            log.info("Time Create Single IFUcube  = %.1f.s" % (t1 - t0,))
+            log.info("Time Create Single ifucube  = %.1f.s" % (t1 - t0,))
 #_______________________________________________________________________
-            single_IFUCube.append(IFUCube)
+            single_ifucube_container.append(ifucube_model)
             del spaxel[:]
-        return single_IFUCube
+        return single_ifucube_container
 #********************************************************************************
     def create_spaxel(self):
         """
@@ -505,7 +941,7 @@ class IFUCubeData():
 
 
                         t1 = time.time()
-                        log.debug("Time Match one Channel from 1 file  to IFUCube = %.1f.s"
+                        log.debug("Time Match one Channel from 1 file to ifucube = %.1f.s"
                                   % (t1 - t0,))
 #________________________________________________________________________________
 #2D area method - only works for single files and coord_system = 'alpha-beta'
@@ -535,7 +971,6 @@ class IFUCubeData():
                             x = x[index]
                             t0 = time.time()
 
-
                             cube_overlap.match_det2cube(self, x, y, i,
                                                         start_region,
                                                         input_model,
@@ -547,7 +982,7 @@ class IFUCubeData():
 #********************************************************************************
                 elif instrument == 'NIRSPEC':
                     # each file, detector has 30 slices - wcs information access seperately for each slice
-                    nslices = 30 
+                    nslices = 30
                     log.info("Mapping each NIRSPEC slice to sky, this takes a while for NIRSPEC data")
                     for i in range(nslices):
 #                        slice_wcs = nirspec.nrs_wcs_set_input(input_model, i)
@@ -565,7 +1000,7 @@ class IFUCubeData():
 
 
                         t1 = time.time()
-                        log.debug("Time Match one NIRSPEC slice  to IFUCube = %.1f.s" % (t1 - t0,))
+                        log.debug("Time Match one NIRSPEC slice to ifucube = %.1f.s" % (t1 - t0,))
 #********************************************************************************
     def find_spaxel_flux(self, spaxel):
 #********************************************************************************
@@ -617,7 +1052,7 @@ class IFUCubeData():
             log.info("Time to interpolate at spaxel values = %.1f.s" % (t1 - t0,))
 
 #********************************************************************************
-    def setup_IFUCube(self, j):
+    def setup_ifucube(self, j):
 
         """
         Short Summary
@@ -645,15 +1080,15 @@ class IFUCubeData():
         dq_cube = np.zeros((naxis3, naxis2, naxis1))
         err_cube = np.zeros((naxis3, naxis2, naxis1))
 
-        IFUCube = datamodels.IFUCubeModel(data=data, dq=dq_cube, err=err_cube, weightmap=idata)
-        IFUCube.update(self.input_models[j])
-        IFUCube.meta.filename = self.output_name
+        ifucube_model = datamodels.IFUCubeModel(data=data, dq=dq_cube, err=err_cube, weightmap=idata)
+        ifucube_model.update(self.input_models[j])
+        ifucube_model.meta.filename = self.output_name
 
         # Call model_blender if there are multiple inputs
         if len(self.input_models) > 1:
-            saved_model_type = IFUCube.meta.model_type
-            self.blend_output_metadata(IFUCube)
-            IFUCube.meta.model_type = saved_model_type  # Reset to original
+            saved_model_type = ifucube_model.meta.model_type
+            self.blend_output_metadata(ifucube_model)
+            ifucube_model.meta.model_type = saved_model_type  # Reset to original
 
 #______________________________________________________________________
         if self.output_type == 'single':
@@ -664,119 +1099,119 @@ class IFUCubeData():
                 self.output_name_base = filename[:indx]
                 self.output_file = None
                 newname = self.define_cubename()
-                IFUCube.meta.filename = newname
+                ifucube_model.meta.filename = newname
 
 #______________________________________________________________________
 # fill in Channel, Band for MIRI
         if self.instrument == 'MIRI':
             # fill in Channel output meta data
             num_ch = len(self.list_par1)
-            IFUCube.meta.instrument.channel = self.list_par1[0]
+            ifucube_model.meta.instrument.channel = self.list_par1[0]
             num_ch = len(self.list_par1)
             for m in range(1, num_ch):
-                IFUCube.meta.instrument.channel = (
-                IFUCube.meta.instrument.channel + str(self.list_par1[m]))
+                ifucube_model.meta.instrument.channel = (
+                ifucube_model.meta.instrument.channel + str(self.list_par1[m]))
 
 #______________________________________________________________________
-        IFUCube.meta.wcsinfo.crval1 = self.Crval1
-        IFUCube.meta.wcsinfo.crval2 = self.Crval2
-        IFUCube.meta.wcsinfo.crval3 = self.Crval3
-        IFUCube.meta.wcsinfo.crpix1 = self.Crpix1
-        IFUCube.meta.wcsinfo.crpix2 = self.Crpix2
-        IFUCube.meta.wcsinfo.crpix3 = self.Crpix3
-        IFUCube.meta.wcsinfo.cdelt1 = self.Cdelt1 / 3600.0
-        IFUCube.meta.wcsinfo.cdelt2 = self.Cdelt2 / 3600.0
-        IFUCube.meta.wcsinfo.cdelt3 = self.Cdelt3
+        ifucube_model.meta.wcsinfo.crval1 = self.Crval1
+        ifucube_model.meta.wcsinfo.crval2 = self.Crval2
+        ifucube_model.meta.wcsinfo.crval3 = self.Crval3
+        ifucube_model.meta.wcsinfo.crpix1 = self.Crpix1
+        ifucube_model.meta.wcsinfo.crpix2 = self.Crpix2
+        ifucube_model.meta.wcsinfo.crpix3 = self.Crpix3
+        ifucube_model.meta.wcsinfo.cdelt1 = self.Cdelt1 / 3600.0
+        ifucube_model.meta.wcsinfo.cdelt2 = self.Cdelt2 / 3600.0
+        ifucube_model.meta.wcsinfo.cdelt3 = self.Cdelt3
 
-        IFUCube.meta.wcsinfo.ctype1 = 'RA---TAN'
-        IFUCube.meta.wcsinfo.ctype2 = 'DEC--TAN'
-        IFUCube.meta.wcsinfo.cunit1 = 'deg'
-        IFUCube.meta.wcsinfo.cunit2 = 'deg'
+        ifucube_model.meta.wcsinfo.ctype1 = 'RA---TAN'
+        ifucube_model.meta.wcsinfo.ctype2 = 'DEC--TAN'
+        ifucube_model.meta.wcsinfo.cunit1 = 'deg'
+        ifucube_model.meta.wcsinfo.cunit2 = 'deg'
 
-        IFUCube.meta.wcsinfo.ctype3 = 'WAVE'
-        IFUCube.meta.wcsinfo.cunit3 = 'um'
-        IFUCube.meta.wcsinfo.wcsaxes = 3
-        IFUCube.meta.wcsinfo.pc1_1 = -1
-        IFUCube.meta.wcsinfo.pc1_2 = 0
-        IFUCube.meta.wcsinfo.pc1_3 = 0
+        ifucube_model.meta.wcsinfo.ctype3 = 'WAVE'
+        ifucube_model.meta.wcsinfo.cunit3 = 'um'
+        ifucube_model.meta.wcsinfo.wcsaxes = 3
+        ifucube_model.meta.wcsinfo.pc1_1 = -1
+        ifucube_model.meta.wcsinfo.pc1_2 = 0
+        ifucube_model.meta.wcsinfo.pc1_3 = 0
 
-        IFUCube.meta.wcsinfo.pc2_1 = 0
-        IFUCube.meta.wcsinfo.pc2_2 = 1
-        IFUCube.meta.wcsinfo.pc2_3 = 0
+        ifucube_model.meta.wcsinfo.pc2_1 = 0
+        ifucube_model.meta.wcsinfo.pc2_2 = 1
+        ifucube_model.meta.wcsinfo.pc2_3 = 0
 
-        IFUCube.meta.wcsinfo.pc3_1 = 0
-        IFUCube.meta.wcsinfo.pc3_2 = 0
-        IFUCube.meta.wcsinfo.pc3_3 = 1
+        ifucube_model.meta.wcsinfo.pc3_1 = 0
+        ifucube_model.meta.wcsinfo.pc3_2 = 0
+        ifucube_model.meta.wcsinfo.pc3_3 = 1
 
-        IFUCube.meta.ifu.flux_extension = 'SCI'
-        IFUCube.meta.ifu.error_extension = 'ERR'
-        IFUCube.meta.ifu.error_type = 'ERR'
-        IFUCube.meta.ifu.dq_extension = 'DQ'
-        IFUCube.meta.ifu.roi_spatial = self.rois
-        IFUCube.meta.ifu.roi_wave = self.roiw
-        IFUCube.meta.ifu.weighting = self.weighting
-        IFUCube.meta.ifu.weight_power = self.weight_power
+        ifucube_model.meta.ifu.flux_extension = 'SCI'
+        ifucube_model.meta.ifu.error_extension = 'ERR'
+        ifucube_model.meta.ifu.error_type = 'ERR'
+        ifucube_model.meta.ifu.dq_extension = 'DQ'
+        ifucube_model.meta.ifu.roi_spatial = self.rois
+        ifucube_model.meta.ifu.roi_wave = self.roiw
+        ifucube_model.meta.ifu.weighting = self.weighting
+        ifucube_model.meta.ifu.weight_power = self.weight_power
 
         with datamodels.open(self.input_models[j]) as input:
-            IFUCube.meta.bunit_data = input.meta.bunit_data
-            IFUCube.meta.bunit_err = input.meta.bunit_err
+            ifucube_model.meta.bunit_data = input.meta.bunit_data
+            ifucube_model.meta.bunit_err = input.meta.bunit_err
 
         if self.coord_system == 'alpha-beta':
-            IFUCube.meta.wcsinfo.cunit1 = 'arcsec'
-            IFUCube.meta.wcsinfo.cunit2 = 'arcsec'
+            ifucube_model.meta.wcsinfo.cunit1 = 'arcsec'
+            ifucube_model.meta.wcsinfo.cunit2 = 'arcsec'
 
 # we only need to check list_par1[0] and list_par2[0] because these types
 # of cubes are made from 1 exposures (setup_cube checks this at the start
 # of cube_build).
             if self.list_par1[0] == '1' and self.list_par2[0] == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL1A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE1A'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL1A'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE1A'
             if self.list_par1[0] == '2' and self.list_par2[0] == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL2A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE2A'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL2A'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE2A'
             if self.list_par1[0] == '3' and self.list_par2[0] == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL3A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE3A'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL3A'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE3A'
             if self.list_par1[0] == '4' and self.list_par2[0] == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL4A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE4A'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL4A'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE4A'
 
             if self.list_par1[0] == '1' and self.list_par2[0] == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL1B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE1B'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL1B'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE1B'
             if self.list_par1[0] == '2' and self.list_par2[0] == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL2B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE2B'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL2B'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE2B'
             if self.list_par1[0] == '3' and self.list_par2[0] == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL3B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE3B'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL3B'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE3B'
             if self.list_par1[0] == '4' and self.list_par2[0] == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL4B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE4B'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL4B'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE4B'
 
             if self.list_par1[0] == '1' and self.list_par2[0] == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL1C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE1C'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL1C'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE1C'
             if self.list_par1[0] == '2' and self.list_par2[0] == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL2C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE2C'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL2C'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE2C'
             if self.list_par1[0] == '3' and self.list_par2[0] == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL3C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE3C'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL3C'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE3C'
             if self.list_par1[0] == '4' and self.list_par2[0] == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL4C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE4C'
+                ifucube_model.meta.wcsinfo.ctype1 = 'MRSAL4C'
+                ifucube_model.meta.wcsinfo.ctype2 = 'MRSBE4C'
 
 
-        wcsobj = pointing.create_fitswcs(IFUCube)
-        IFUCube.meta.wcs = wcsobj
-        IFUCube.meta.wcs.bounding_box = ((0, naxis1 - 1),
+        wcsobj = pointing.create_fitswcs(ifucube_model)
+        ifucube_model.meta.wcs = wcsobj
+        ifucube_model.meta.wcs.bounding_box = ((0, naxis1 - 1),
                                          (0, naxis2 - 1),
                                          (0, naxis3 - 1))
-        return IFUCube
+        return ifucube_model
 
 #********************************************************************************
-    def update_IFUCube(self, IFUCube, spaxel):
+    def update_ifucube(self, ifucube_model, spaxel):
 #********************************************************************************
         """
         Short Summary
@@ -801,9 +1236,9 @@ class IFUCubeData():
                           [self.naxis3, self.naxis2, self.naxis1])
 
 
-        IFUCube.data = temp_flux
-        IFUCube.weightmap = temp_wmap
-        IFUCube.meta.cal_step.cube_build = 'COMPLETE'
+        ifucube_model.data = temp_flux
+        ifucube_model.weightmap = temp_wmap
+        ifucube_model.meta.cal_step.cube_build = 'COMPLETE'
 #    icube = 0
 #    for z in range(Cube.naxis3):
 #        for y in range(Cube.naxis2):
@@ -811,27 +1246,6 @@ class IFUCubeData():
 #                IFUCube.data[z, y, x] = spaxel[icube].flux
 #                IFUCube.weightmap[z, y, x] = len(spaxel[icube].ipointcloud)
 #                icube = icube + 1
-
-
-       # result = IFUCube.copy()
-        #return result
-
-#********************************************************************************
-#********************************************************************************
-#    def find_output_type(self):
-
-#        ValidChannel = ['1', '2', '3', '4']
-#        ValidSubchannel = ['SHORT', 'MEDIUM', 'LONG']
-
-#        nchannels = len(ValidChannel)
-#        nsubchannels = len(ValidSubchannel)
-
-#        ValidGWA = ['G140M', 'G140H', 'G140M', 'G140H', 'G235M', 'G235H',
-#                    'G395M', 'G395H', 'PRISM']
-#        ValidFWA = ['F070LP', 'F070LP', 'F100LP', 'F100LP', 'F170LP',
-#                    'F170LP', 'F290LP', 'F290LP', 'CLEAR']
-
-#        nbands = len(ValidFWA)
 
 #********************************************************************************
     def blend_output_metadata(self, IFUCube):
